refactor account api; rename snsetters -> starknet; delete sn_models; delete sn_getters
